edit art details as admin

+++ art/models.py
@@ -39,7 +39,7 @@ class Art(models.Model):
     description = models.TextField()
     meaning = models.TextField()
     color = models.TextField()
-    price = models.IntegerField()
+    price = models.DecimalField(max_digits=6, decimal_places=0)
     date = models.DateTimeField(auto_now_add=True)
     image_url = models.URLField(max_length=1024, null=True, blank=True)
     image = models.ImageField(null=True, blank=True)
